* stable: (24 commits)
  (GH-839) Switch to apply package parameters to dependent packages
  (maint) formatting methods / parameters in calls
  (GH-958) If SSLv3 in Posh v2 Fails, Use Original
  (GH-746) Use HTTPS if available when HTTP url
  (GH-957) Skip Get-WebFileName When FTP
  (GH-948) Ensure passwords / keys are not logged
  (GH-952) Get-ChocolateyWebFile enhancements
  (doc) update generated docs
  (maint) formatting
  (docs) move GenerateDocs.ps1 / update
  (GH-932) Do not set unknown checksum to 'md5'
  (GH-719) Reset ServerCertificateValidationCallback
  (GH-305) add MSP/MSU installer types
  (GH-305) update exitcodes to long
  (GH-954) Pending fails when lib does not exist
  (GH-950) Install-ChocolateyPackage - UseOriginalLocation
  (maint) formatting
  (GH-922) Automatically determine checksum type
  (maint) fixes for shimgen
  (GH-948) Do not log sensitive arguments
  ...

+++ src/chocolatey/infrastructure.app/domain/installers/GhostInstaller.cs
@@ -39,14 +39,8 @@ namespace chocolatey.infrastructure.app.domain.installers
             UninstallExecutable = "\"{0}\"".format_with(InstallTokens.UNINSTALLER_LOCATION);
             SilentUninstall = "-u -s";
             OtherUninstallOptions = "";
-            ValidInstallExitCodes = new List<int>
-            {
-                0
-            };
-            ValidUninstallExitCodes = new List<int>
-            {
-                0
-            };
+            ValidInstallExitCodes = new List<long> { 0 };
+            ValidUninstallExitCodes = new List<long> { 0 };
         }
 
         public override InstallerType InstallerType { get { return InstallerType.Ghost; } }
